Rename package name "perfectfit" to "perfectfitapp"
